# iam-management-web/src/main/resources/openapi/iam.yaml
openapi: 3.0.3
info:
  title: Generic IAM Management API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /iam/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentPrincipal
      responses:
        '200':
          description: Current principal
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PrincipalResponse'}
        '401': {description: Authentication required}
  /iam/sessions:
    get:
      tags: [Sessions]
      operationId: listMySessions
      parameters:
        - name: userId
          in: query
          required: false
          deprecated: true
          schema: {type: integer, format: int64}
          description: Ignored compatibility parameter; identity always comes from the Bearer principal.
      responses:
        '200':
          description: Active sessions owned by the current principal
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionListResponse'}
        '401': {description: Authentication required}
  /iam/sessions/{sessionId}/revoke:
    post:
      tags: [Sessions]
      operationId: revokeMySession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RevokeRequest'}
      responses:
        '204': {description: Session revoked}
        '401': {description: Authentication required}
        '404': {description: Active session not found for current principal}
  /iam/sessions/revoke-others:
    post:
      tags: [Sessions]
      operationId: revokeOtherSessions
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RevokeOthersRequest'}
      responses:
        '204': {description: Other sessions revoked}
        '401': {description: Authentication required}
...
  /iam/authorization/profiles/{profileId}/switch:
    post:
      tags: [Authorization]
      operationId: switchAuthorizationProfile
      parameters:
        - name: profileId
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        '200':
          description: New token with exactly one active profile
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginResponse'}
        '401':
          description: Authentication required
        '404':
          description: Profile is not available to the current principal
  /iam/authorization/diagnostics:
    post:
      tags: [Authorization]
      operationId: evaluateAuthorization
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthorizationEvaluationRequest'}
      responses:
        '200':
          description: Projection of the runtime authorization decision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthorizationDecisionResponse'}
...
components:
  schemas:
...
    LoginResponse:
      type: object
      required: [accessToken, sessionId, expiresAt, principal]
      properties:
        accessToken: {type: string}
        sessionId: {type: string}
        expiresAt: {type: string, format: date-time}
        principal: {$ref: '#/components/schemas/PrincipalResponse'}
    PrincipalResponse:
      type: object
      required: [userId, identityId, identityDomain, clientType, authorizationVersion]
      properties:
        userId: {type: integer, format: int64}
        identityId: {type: string}
        identityDomain: {type: string}
        activeProfileId: {type: integer, format: int64, nullable: true}
        templateVersionId: {type: integer, format: int64, nullable: true}
        clientType: {type: string}
        authorizationVersion: {type: integer, format: int64}
    SessionResponse:
      type: object
      required: [sessionId, userId, clientType, loginAt, lastSeenAt, expiresAt]
      properties:
        sessionId: {type: string}
        userId: {type: integer, format: int64}
        clientType: {type: string}
        clientInstance: {type: string, nullable: true}
        ipAddress: {type: string, nullable: true}
        userAgent: {type: string, nullable: true}
        loginAt: {type: string, format: date-time}
        lastSeenAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
        logoutAt: {type: string, format: date-time, nullable: true}
    SessionListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/SessionResponse'}
    RevokeRequest:
      type: object
      properties:
        reason: {type: string, default: USER_REQUEST}
    RevokeOthersRequest:
      type: object
      required: [currentSessionId]
      properties:
        currentSessionId: {type: string, minLength: 1}
        reason: {type: string, default: USER_REQUEST}
...
    AuthorizationEvaluationRequest:
      type: object
      required: [permissionCode, domain, clientType, resourceType, resourceId, scopeAccess]
      properties:
        permissionCode: {type: string}
        domain: {type: string}
        clientType: {type: string}
        resourceType: {type: string}
        resourceId: {type: string}
        scopeAccess: {type: string, enum: [READ, WRITE]}
    AuthorizationDecisionStep:
      type: object
      required: [code, passed, reason]
      properties:
        code: {type: string}
        passed: {type: boolean}
        reason: {type: string}
    AuthorizationDecisionResponse:
      type: object
      required: [allowed, decisionCode, steps]
      properties:
        allowed: {type: boolean}
        decisionCode: {type: string}
        steps: {type: array, items: {$ref: '#/components/schemas/AuthorizationDecisionStep'}}